order files and clean unused
all scripts that a user would execute go to .tel/bin. others go to ./tel/scripts with the right file ending. config files that the user could edit go to ./tel/configs. any files that are used by scripts but are usually not modified by the user go to .tel/data